+++ CHANGES.md
@@ -1,4 +1,7 @@
 # [Unreleased]
+- ...
+
+# [0.3] - 2025-11-26
 - (!backwards incompatible) change of attribute `showIcon` ("no, "before", "after") to `iconPlacement` ("before", "after"). affected components: `button`, `link`, `tag`. The presence of the `icon` attribute is now enough to show (default "before")
 - add customization options `css_class_mappings` and `attribute_removals`
 - fix that alert component shows correct icon: different per kind